+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPQZ9EWC7Q8B6168X6M67Q8Q"
+        "01KPR03Y7HSA6SMF36TQJTV1KQ"
     )
     r"""The unique identifier of the data source"""