+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KPQZ9EW9C0FHPWF6VC6QKH1C"
+        "01KPR03Y7FMA7QYD85ARJ1NR1K"
     )
     r"""The unique identifier of the data source"""