+++ packages/orq-rc/RELEASES.md
@@ -7159,3 +7159,13 @@ Based on:
 - [python v4.8.0-rc.38] packages/orq-rc
 ### Releases
 - [PyPI v4.8.0-rc.38] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.38 - packages/orq-rc
+
+## 2026-04-21 12:29:43
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.9 (2.881.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.8.0-rc.39] packages/orq-rc
+### Releases
+- [PyPI v4.8.0-rc.39] https://pypi.org/project/orq-ai-sdk/4.8.0-rc.39 - packages/orq-rc